refactored some code to clean things up, extracted out some common logic
