+++ src/lambda/populate_provider_versions/main.go
@@ -5,16 +5,9 @@ import (
 	"encoding/json"
 	"fmt"
 	"github.com/aws/aws-lambda-go/lambda"
-	"github.com/aws/aws-sdk-go-v2/aws"
-	awsconfig "github.com/aws/aws-sdk-go-v2/config"
-	"github.com/aws/aws-sdk-go-v2/service/secretsmanager"
-	"github.com/aws/aws-xray-sdk-go/instrumentation/awsv2"
 	"github.com/aws/aws-xray-sdk-go/xray"
-	gogithub "github.com/google/go-github/v54/github"
 	"github.com/opentffoundation/registry/internal/github"
 	"github.com/opentffoundation/registry/internal/providers"
-	"github.com/shurcooL/githubv4"
-	"os"
 )
 
 type PopulateProviderVersionsEvent struct {
@@ -22,145 +15,77 @@ type PopulateProviderVersionsEvent struct {
 	Type      string `json:"type"`
 }
 
-type Config struct {
-	ManagedGithubClient *gogithub.Client
-	RawGithubv4Client   *githubv4.Client
-	ProviderRedirects   map[string]string
+func (p PopulateProviderVersionsEvent) Validate() error {
+	if p.Namespace == "" {
+		return fmt.Errorf("namespace is required")
+	}
+	if p.Type == "" {
+		return fmt.Errorf("type is required")
+	}
+	return nil
 }
 
-var config *Config
-
-func init() {
+func main() {
 	ctx := context.Background()
-
-	c, err := buildConfig(ctx)
+	config, err := buildConfig(ctx)
 	if err != nil {
 		panic(fmt.Errorf("could not build config: %w", err))
 	}
 
-	config = c
+	lambda.Start(HandleRequest(config))
 }
 
-func buildConfig(ctx context.Context) (config *Config, err error) {
-	if err = xray.Configure(xray.Config{ServiceVersion: "1.2.3"}); err != nil {
-		err = fmt.Errorf("could not configure X-Ray: %w", err)
-		return
-	}
+func HandleRequest(config *Config) func(ctx context.Context, e PopulateProviderVersionsEvent) (string, error) {
+	return func(ctx context.Context, e PopulateProviderVersionsEvent) (string, error) {
+		var versions []providers.Version
 
-	// fetch the github token ASM name from the environment
-	githubTokenSecretName := os.Getenv("GITHUB_TOKEN_SECRET_ASM_NAME")
-	if githubTokenSecretName == "" {
-		panic("GITHUB_TOKEN_SECRET_ASM_NAME environment variable not set")
-	}
+		fmt.Printf("Fetching %s/%s\n", e.Namespace, e.Type)
+		err := xray.Capture(ctx, "populate_provider_versions.handle", func(tracedCtx context.Context) error {
+			xray.AddAnnotation(tracedCtx, "namespace", e.Namespace)
+			xray.AddAnnotation(tracedCtx, "type", e.Type)
 
-	// At this point we're not part of a Lambda request execution, so let's
-	// explicitly create a segment to represent the configuration process.
-	ctx, segment := xray.BeginSegment(ctx, "registry.config")
-	defer func() { segment.Close(err) }()
+			err := e.Validate()
+			if err != nil {
+				return fmt.Errorf("invalid event: %w", err)
+			}
 
-	var secretsmanager *secretsmanager.Client
-	if secretsmanager, err = getSecretsManager(ctx); err != nil {
-		err = fmt.Errorf("could not get secrets manager client: %w", err)
-		return
-	}
+			// Construct the repo name.
+			repoName := providers.GetRepoName(e.Type)
 
-	var githubAPIToken string
-	if githubAPIToken, err = getSecretValue(ctx, secretsmanager, githubTokenSecretName); err != nil {
-		err = fmt.Errorf("could not get GitHub API token: %w", err)
-		return
-	}
+			// check the repo exists
+			exists, err := github.RepositoryExists(ctx, config.ManagedGithubClient, e.Namespace, repoName)
+			if err != nil {
+				return fmt.Errorf("failed to check if repo exists: %w", err)
+			}
+			if !exists {
+				return fmt.Errorf("repo does not exist")
+			}
 
-	if githubAPIToken == "" {
-		err = fmt.Errorf("empty GitHub API token fetched from secrets manager")
-		return
-	}
+			fmt.Printf("Repo %s/%s exists\n", e.Namespace, repoName)
 
-	fmt.Printf("GitHub API token: %s\n", githubAPIToken)
+			v, err := providers.GetVersions(tracedCtx, config.RawGithubv4Client, e.Namespace, repoName)
+			if err != nil {
+				return fmt.Errorf("failed to get versions: %w", err)
+			}
 
-	config = &Config{
-		ManagedGithubClient: github.NewManagedGithubClient(githubAPIToken),
-		RawGithubv4Client:   github.NewRawGithubv4Client(githubAPIToken),
-	}
+			fmt.Printf("Found %d versions\n", len(v))
 
-	return
-}
+			versions = v
+			return nil
+		})
 
-func getSecretsManager(ctx context.Context) (*secretsmanager.Client, error) {
-	awsConfig, err := awsconfig.LoadDefaultConfig(ctx, awsconfig.WithRegion(os.Getenv("AWS_REGION")))
-	if err != nil {
-		return nil, fmt.Errorf("could not load AWS configuration: %w", err)
-	}
-
-	awsv2.AWSV2Instrumentor(&awsConfig.APIOptions)
-
-	return secretsmanager.NewFromConfig(awsConfig), nil
-}
-
-func getSecretValue(ctx context.Context, sm *secretsmanager.Client, secretName string) (string, error) {
-	value, err := sm.GetSecretValue(ctx, &secretsmanager.GetSecretValueInput{
-		SecretId: aws.String(secretName),
-	})
-	if err != nil {
-		return "", err
-	}
-	return *value.SecretString, nil
-}
-
-func main() {
-	lambda.Start(HandleRequest)
-}
-
-func HandleRequest(ctx context.Context, e PopulateProviderVersionsEvent) (string, error) {
-	var versions []providers.Version
-
-	fmt.Printf("Fetching %s/%s\n", e.Namespace, e.Type)
-	err := xray.Capture(ctx, "populate_provider_versions.handle", func(tracedCtx context.Context) error {
-		xray.AddAnnotation(tracedCtx, "namespace", e.Namespace)
-		xray.AddAnnotation(tracedCtx, "type", e.Type)
-
-		if e.Namespace == "" {
-			return fmt.Errorf("namespace is required")
-		}
-		if e.Type == "" {
-			return fmt.Errorf("type is required")
-		}
-
-		// Construct the repo name.
-		repoName := providers.GetRepoName(e.Type)
-
-		// check the repo exists
-		exists, err := github.RepositoryExists(ctx, config.ManagedGithubClient, e.Namespace, repoName)
 		if err != nil {
-			return fmt.Errorf("failed to check if repo exists: %w", err)
-		}
-		if !exists {
-			return fmt.Errorf("repo does not exist")
+			fmt.Printf("error: %s\n", err.Error())
+			return "", err
 		}
 
-		fmt.Printf("Repo %s/%s exists\n", e.Namespace, repoName)
+		// TODO: Send to dynamodb
 
-		v, err := providers.GetVersions(tracedCtx, config.RawGithubv4Client, e.Namespace, repoName)
+		marshalled, err := json.Marshal(versions)
 		if err != nil {
-			return fmt.Errorf("failed to get versions: %w", err)
+			return "", fmt.Errorf("failed to marshal versions: %w", err)
 		}
 
-		fmt.Printf("Found %d versions\n", len(v))
-
-		versions = v
-		return nil
-	})
-
-	if err != nil {
-		fmt.Printf("error: %s\n", err.Error())
-		return "", err
+		return string(marshalled), nil
 	}
-
-	// TODO: Send to dynamodb
-
-	marshalled, err := json.Marshal(versions)
-	if err != nil {
-		return "", fmt.Errorf("failed to marshal versions: %w", err)
-	}
-
-	return string(marshalled), nil
 }
